--- a/src/gsl/pointers.h
+++ b/src/gsl/pointers.h
@@ -20,12 +20,10 @@
 #include <gsl/assert.h> // for Ensures, Expects
 #include <source_location.h>
 
-#include <algorithm>    // for forward
 #include <cstddef>      // for ptrdiff_t, nullptr_t, size_t
-#include <memory>       // for shared_ptr, unique_ptr
-#include <system_error> // for hash
+#include <memory>       // for shared_ptr, unique_ptr, hash
 #include <type_traits>  // for enable_if_t, is_convertible, is_assignable
-#include <utility>      // for declval
+#include <utility>      // for declval, forward
 
 #if !defined(GSL_NO_IOSTREAMS)
 #include <iosfwd> // for ostream