@knst 2024-06-03
> Let's wait for the upstream PR to be merged first

for reference: https://github.com/microsoft/GSL/pull/1153