@knst 2024-08-26
@PastaPastaPasta @UdjinM6 let's get this one merged. I compile Dash every day on laptop and extra 10 seconds off compilation time for clean build is a noticeable improvement for me.

I will create follow-up PR once that one is merged (if any difference to merge): https://github.com/microsoft/GSL/pull/1153